app/requirements.txt
boto3
Flask
flask_cors
python-dotenv
requests